just a safety check so we don't kill another process with the same PID if socat went away, as it does when the connection is refused - I can't easily detect a connection failure right now

+++ src/ProbeNetDevice/ProbeNetDevice.sh
@@ -192,7 +192,7 @@ DisconnectFromRemoteHost()
 		exec 3>&-
 	fi
 
-	if [[ -n "$PID_RemoteHostSocat" ]]; then
+	if [[ -n "$PID_RemoteHostSocat" && -d "/proc/$PID_RemoteHostSocat" ]] && grep -qF "$0" "/proc/$PID_RemoteHostSocat/cmdline"; then
 		kill "$PID_RemoteHostSocat"
 	fi
 }
